Update G711 PLC

+++ plc/plc/test.c
@@ -213,7 +213,7 @@ int read_wav_8khz(int16_t* output, int max_samples, int16_t* raw_pcm_buf, int16_
         return 0;
     // 读取原始PCM数据（复用预分配缓冲区）
     size_t bytes_read = fread(raw_pcm_buf, sizeof(int16_t), samples_needed * wav_channels, wav_file);
-    if (bytes_read < samples_needed * wav_channels) {
+    if ((int)bytes_read < samples_needed * wav_channels) {
         return 0;
     }
     int frames_read = bytes_read / wav_channels;
